+++ shaders/common/settings.glsl
@@ -144,6 +144,12 @@ const float shadowIntervalSize    = 2.0;
 
 
 
+#if (defined OVERWORLD && CUSTOM_OVERWORLD_SKYBOX == 1) || (defined NETHER && CUSTOM_NETHER_SKYBOX == 1) || (defined END && CUSTOM_END_SKYBOX == 1)
+	#define CUSTOM_SKYBOX
+#endif
+
+
+
 // How on earth can optifine be THIS bad???
 #if TEMPORAL_FILTER_ENABLED == 1
 	#undef TEMPORAL_FILTER_ENABLED