+++ shaders/style_fantasy.glsl
@@ -1168,7 +1168,7 @@
 #endif
 #if CLOUD_SUNRISE_BLUE == -1
 #undef CLOUD_SUNRISE_BLUE
-										#define CLOUD_SUNRISE_BLUE 0.35
+										#define CLOUD_SUNRISE_BLUE 0.4
 #endif
 #if CLOUD_SUNSET_RED == -1
 #undef CLOUD_SUNSET_RED