+++ shaders/program/clrwl_shadow.glsl
@@ -29,7 +29,7 @@ void main() {
 	
 	texcoord = (gl_TextureMatrix[0] * gl_MultiTexCoord0).xy;
 	
-	#if COLORED_LIGHTING_ENABLED == 1 || EXCLUDE_FOLIAGE == 1 || WAVING_ENABLED == 1 || PHYSICALLY_WAVING_WATER_ENABLED == 1
+	#if COLORED_LIGHTING_ENABLED == 1
 		uint materialId = uint(max(int(mc_Entity.x) - 10000, 0));
 	#endif